Made the script robust against files with multiple key cycles
In ROOT files, a certain key may have multiple "cycles" associated with
it. These are multiple versions of the same data, and generally appear
as a result of overwriting an already existing file, or randomly as a
side-effect of ROOT's auto-save feature.

Until now, we did not manage this information correctly. Now we do.