Cleaned up RandomNumbersSvc
Back when I introduced thread-safe and reproducible Alg-local random
number generators in the test framework, I chose to keep around a
"global" random number generator in RandomNumbersSvc for compatibility
and simple use cases.

This turned out to be a mistake: it confused people, and it was never
used as intended. So I propose to remove this facility, and make Alg-local
generators the only supported way of generating random numbers in the
ACTS test framework.

I also added some extra documentation to clarify usage of
RandomNumbersSvc, and removed the unused pileupType configuration
parameter in the ReadEvgen algorithm, which gives the misleading
impression that one can choose the random numbers distribution for
pileup (it is actually always Poissonian in the code).