Removed LSP-violating string-based writer methods
Some writer interfaces have a write() overload which takes an
"additional string to be written out". This is overly restrictive as
only text writers can emit this output, forcing other writers to ignore
it silently which is a Bad Thing.

In practice, this interface method was only used by
ObjTrackingGeometryWriter. So I specialized it so that it doesn't need
the method to be part of IWriterT anymore, and am now dropping the
method.